Move requests to the `peer` module
